fix: resolve circular import by moving capture_populate_pod_logs to mtv_migration
The tox check was failing due to a circular import:
- conftest.py imports from utilities.mtv_migration
- utilities.mtv_migration imports capture_populate_pod_logs from utilities.copyoffload_migration
- utilities.copyoffload_migration imports get_migration_for_plan and wait_for_migration_complate from utilities.mtv_migration
- Result: ImportError during pytest collection

Solution:
- Moved capture_populate_pod_logs() from copyoffload_migration.py to mtv_migration.py
- This function is only called from mtv_migration.py, so it belongs there
- Added necessary imports to mtv_migration.py: Pod and PVC_NAME_LABEL
- Removed the function from copyoffload_migration.py

This breaks the circular dependency:
- mtv_migration.py now has capture_populate_pod_logs locally (no import needed)
- copyoffload_migration.py still imports from mtv_migration.py (one-way dependency)
- No circular import!

Verified:
- pytest --collect-only: 244 tests collected successfully
- All pre-commit checks pass (mypy, flake8, ruff, formatting)

Fixes the tox failure that was blocking the PR.
